Re-organized location into sub package and added path and feed.
