+++ client/voyager/location/__init__.py
@@ -1,0 +1,4 @@
+from location import Location
+from data import *
+from feed import *
+from path import *
